feature/basic-unit-test
set up basic unit test for all components
